requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
python-multipart
pydantic
pydantic[email]
pandas
requests
beautifulsoup4
apscheduler
passlib[bcrypt]
python-jose[cryptography]
groq
openai
selenium
webdriver-manager
openpyxl
xlrd
werkzeug
PyJWT
reportlab
websockets
elevenlabs